create_post_work: Refactored app structure. Implement and use express router instead of using express directly to handle requests. Add API to create a new post.
